Make FilePreparation Step to work and update configuration

+++ src/main/java/example/billingapp/FilePreparationTasklet.java
@@ -15,12 +15,23 @@ public class FilePreparationTasklet implements Tasklet {
 
     @Override
     public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) throws Exception {
-        JobParameters jobParameters = chunkContext.getStepContext().getStepExecution().getJobParameters();
+        JobParameters jobParameters = contribution.getStepExecution().getJobParameters();
         String inputFile = jobParameters.getString("input.file");
-        Path source = Paths.get(inputFile);
-        Path target = Paths.get("staging", source.toFile().getName());
-        Files.copy(source, target, StandardCopyOption.REPLACE_EXISTING);
 
-        return RepeatStatus.FINISHED;
+        if (inputFile != null) {
+            Path source = Paths.get(inputFile);
+            Path target = Paths.get("staging", source.toFile().getName());
+
+            // Ensure the source file exists before copying
+            if (Files.exists(source)) {
+                Files.copy(source, target, StandardCopyOption.REPLACE_EXISTING);
+                System.out.println("File " + source + " copied to: " + target.toAbsolutePath());
+                return RepeatStatus.FINISHED;
+            } else {
+                throw new IllegalArgumentException("Source file does not exist: " + inputFile);
+            }
+        } else {
+            throw new IllegalArgumentException("Input file is null");
+        }
     }
 }
